search_screen 위젯 분리 - 오류 발생 - routes에서 search에 존재하는 SectionTitlePadding의 클래스 명이 겹침 - routes에서는 RouteListCard만 필요하므로 show RouteListCard를 사용하여 오류 해결
