+++ lib/features/search/widgets/search_field.dart
@@ -1,0 +1,40 @@
+import 'package:flutter/material.dart';
+
+// AppBar에 들어가는 검색 입력 필드
+class SearchField extends StatelessWidget {
+  final TextEditingController controller;
+  final ValueChanged<String> onSubmitted;
+
+  const SearchField({
+    required this.controller,
+    required this.onSubmitted,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    final cs = Theme.of(context).colorScheme;
+    return TextField(
+      controller: controller,
+      autofocus: true,
+      textInputAction: TextInputAction.search,
+      onSubmitted: onSubmitted,
+      decoration: InputDecoration(
+        hintText: "루트 이름, 태그, 위치로 검색",
+        isDense: true,
+        filled: true,
+        fillColor: cs.surfaceContainerHighest,
+        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(12),
+          borderSide: BorderSide(color: cs.outlineVariant),
+        ),
+        prefixIcon: const Icon(Icons.search_rounded),
+        suffixIcon: IconButton(
+          tooltip: "지우기",
+          icon: const Icon(Icons.close_rounded),
+          onPressed: () => controller.clear(),
+        ),
+      ),
+    );
+  }
+}
