+++ lib/features/search/widgets/section_title.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class SectionTitlePadding extends StatelessWidget {
+  final String text;
+  const SectionTitlePadding(this.text);
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.fromLTRB(16, 0, 16, 8),
+      child: Text(
+        text,
+        style: Theme.of(context).textTheme.titleMedium?.copyWith(
+          fontWeight: FontWeight.w700,
+        ),
+      ),
+    );
+  }
+}
