+++ lib/features/search/widgets/widgets.dart
@@ -1,0 +1,3 @@
+export 'search_field.dart';
+export 'section_title.dart';
+
